backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
httpx>=0.27.0,<1
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
anthropic>=0.39.0
cachetools==5.5.0
pytest==8.3.4
pytest-asyncio==0.24.0
